Move settings related files into one directory
